#1 Pages created on the right locale
Pages must be created on the right locale or they cannot be found later.

Besides, although we are visiting an spanish route, I18n.locale is equal to :en and the url was always printer in english. This must be some weird behaviour in Rspec and Refinery as it works properly when the app is browsed in development mode...